backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.1.0
whitenoise>=6.4.0
python-dotenv>=1.0.0
Pillow>=9.5.0
drf-yasg>=1.21.5
gunicorn>=21.2.0
pathspec==0.11.2
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0 